requirements.txt
appdirs==1.4.3
asgiref==3.3.1
Babel==2.9.1
boto3==1.17.61
botocore==1.20.61
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.1.6
django-liqpay==1.1
django-mathfilters==1.0.0
django-phone-field==1.8.1
django-phonenumber-field==5.1.0
django-recaptcha==2.0.6
django-storages==1.11.1
django-telegram-login==0.2.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jmespath==0.10.0
liqpay==1.0.1
liqpay-python==1.0
liqpay-sdk-python3==1.0.3
lockfile==0.12.2
mongoengine==0.23.1
msgpack==0.6.2
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.22
Pillow==8.1.0
pkg-resources==0.0.0
progress==1.5
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-http-client==3.3.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.4.2
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.1.0
social-auth-storage-mongoengine==1.0.1
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1